Remove non-standard SCRAM-SHA-512 auth mechanism
While looking into Jackal's auth code I noticed that it supports a SASL
mechanism called SCRAM-SHA-512. There is no such mechanism currently
standardized by the IETF, and it is not a mechanism supported by any
XMPP clients that I am aware of. Because the hash is only used as part
of an HMAC, there are no known security issues at this time with
SCRAM-SHA-1 or SCRAM-SHA-256, and no known benefit to using SHA-512 as
the hash. Furthermore, if the IETF were to standardize a mechanism
called SCRAM-SHA-512 in the future and it were somehow incompatible with
this one, it could be hard to ensure the two were compatible later and
Jackal may be stuck with no upgrade path and no way to tell which
mechanism a client was requesting.

Fixes #107