Remove the util/rand package
A common signal that Go code is not idiomatic is "tiny package
syndrome": a project with lots of packages that usually have a single
file, or a small amount of code that could easily be moved elsewhere.
This sort of over taxonification can make projects hard to navigate and
understand. Since the single function in the rand package only saved one
or two lines, I have taken this opportunity to inline it since I was
working on that part of the code anyways.