Remove DIGEST-MD5 auth
The DIGEST-MD5 auth mechanism has been obsoleted and broken for quite
some time. It would be essentially the same as doing PLAIN auth except
that it also prevents us from storing the password hashed in the
database (because we need to keep the plain password around for MD5
hashing). Legacy clients that support DIGEST-MD5 are likely to support
PLAIN (which still allows us to store a hashed password in the database)
as well, so it is not likely that removing DIGEST-MD5 will create any
compatibility issues, but it will enable a massive security benefit.

See also #62