Store salted and hashed passwords in the database
Because Jackal supports the SCRAM-SHA-1 and SCRAM-SHA-256 authentication
mechanisms, we are already using the PBKDF2 key derivation function on
passwords during auth. This patch updates the database to store the
salted and hashed password, as well as the inputs to PBKDF2 instead of
the plaintext password. It also updates other authentication mechanisms
such as PLAIN, and the registration feature to hash the password using
PBKDF2 and compare with the hashed password in the database.

Because SCRAM-SHA-1 and SCRAM-SHA-256 use different hashes in their
HMAC, unfortunately their salted passwords are not compatible and if we
want to support users who may have multiple clients that don't support
both we have to store salted passwords derived with both. This is a much
more expensive operation, but ensures maximum compatibility with clients
in the wild (for example, Conversations could use SCRAM-SHA-256, but the
user would still be able to log in from their desktop with a client that
only supports SCRAM-SHA-1).

Fixes #62