Fix broken SCRAM auth
While attempting to log in using a client that supports both SCRAM-SHA-1
and SCRAM-SHA-1-PLUS I noticed that login was failing unless I disabled
every mechanism except PLAIN. After futher investigation, I noticed that
the client was using TLS 1.3, which does not support the tls-unique
mechanism used by Jackals channel binding implementation. This would
also have been the case if I had been trying to do TLS resumption
without the TLS master-secret fix.

To fix this I altered the channel binding code to only advertise the
-PLUS variants if the TLS unique data is available. However, even with
the fix login was still failing. It turns out there was also a bug where
the value of the GS2 header was treating clients that report that
channel binding is supported, but not required, as if it was supported
and required. I also added a fix for this, and a comment pointing at the
definitions of the channel binding headers.