Rework Makefile to fix bugs and respect LDFLAGS
Previously every rule in the makefile was phony, but none were declared
as such. This meant that eg. if a file called "clean" or "build" was
created, Make would think the rules with the same names referred to
those files and may think they were up to date and refuse to run the
recipe. Marking these rules as phony ensures that they are always
considered out of date and will always be run.

This also adds the .POSIX and .SUFFIXES targets to remove pre-defined C
suffix rules (its not likely that this would ever cause problems, but
it's nice to start from a clean slate) and use POSIX compatible behavior
(GNU Make is particularly bad about this, so using .POSIX normalizes the
behavior a tiny bit between GNU Make and NetBSD make like I use).

Finally, the build rule has been marked phony and now just depends on
the jackal (non-phony) rule which ensures that go.mod gets run when
necessary, and is rebuilt any time a *.go file changes. The build has
also been updated to trim paths so you don't wind up with release
binaries that include your home directory path in them for debugging
purposes and respect external LDFLAGS to make it easier for package
managers to spit out a position independent executable, or make other
changes that might be required by their operating system or distros
packaging procedures.