scripts: do not install utilities on users system
Previously the genproto script would pollute the users $GOBIN with a
specific version of two tools for generating protobuf files, overwriting
the users existing versions if they already had these tools installed.
This patch depends on those tools in go.mod so that they can be built,
downloaded, and used with go run and don't need to be installed on the
users system. As a bonus, all versioning information is now in go.mod
instead of spread over go.mod and a script.

Signed-off-by: Sam Whited <sam@samwhited.com>